cleaned up code Refactored classes/functions
